Change the structure of ParsedArguments and update parsing functions accordingly
This one is a huge quality of life improvement (in my opinion). Parsing the object returned by yargs-parser was absolutely hell, since Object.entries had to be used. The code got rapidly harder and harder to read and I decided to change the structure of the parsed arguments using an interface, and it made every parsing function a bit easier to read and debug in my opinion.