Migrate to typescript (#8)
* Add typescript

* Create interfaces and refactor available flags as array function

* Add ParsedArguments interface and add types to get matching flags function

* Refactor the get parsed descriptions function to work with TS and change the target to ESNext in tsconfig

* Edit ParsedArguments interface and add SpecialTokens interface

* Update a few functions to work with typescript

* Remove webvitals

* Update the utils functions to work with typescript

* Change the structure of ParsedArguments and update parsing functions accordingly

This one is a huge quality of life improvement (in my opinion). Parsing the object returned by yargs-parser was absolutely hell, since Object.entries had to be used. The code got rapidly harder and harder to read and I decided to change the structure of the parsed arguments using an interface, and it made every parsing function a bit easier to read and debug in my opinion.

* Update app.tsx to work with typescript

This is a huge commit but it is actually way simpler than it looks; mostly function refactors to comply with Typescript, and I also rewrote one function, replaceSpecialTokens, because it did way more than it should and replaced the entire object.

* Fix flag arguments parsing function not working anymore

* Update get parsed flags descriptions to work with multiple times the same flag (needs a better implementation)